requirements.txt
numpy==1.20.0
gym==0.22.0
gym[atari]==0.22.0
gym[accept-rom-license]==0.22.0
cython
PyOpenGL-accelerate
hydra-core
protobuf==3.20.1
wandb==0.14.2
dm_control
opencv-python
kornia
imageio==2.31.5
imageio[ffmpeg]
imageio[pyav]
ray==1.0.0
torchrl==0.1.0
git+https://gitclone.com/github.com/denisyarats/dmc2gym.git
